Move GroupSpinner to ui module
